+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/prepayment/ChangePaymentModeResponse.java
@@ -184,10 +184,10 @@ public class ChangePaymentModeResponse extends ZclPrepaymentCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        friendlyCredit = (Integer) deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
-        friendlyCreditCalendarId = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_32_BIT_INTEGER);
-        emergencyCreditLimit = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_32_BIT_INTEGER);
-        emergencyCreditThreshold = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_32_BIT_INTEGER);
+        friendlyCredit = deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
+        friendlyCreditCalendarId = deserializer.deserialize(ZclDataType.UNSIGNED_32_BIT_INTEGER);
+        emergencyCreditLimit = deserializer.deserialize(ZclDataType.UNSIGNED_32_BIT_INTEGER);
+        emergencyCreditThreshold = deserializer.deserialize(ZclDataType.UNSIGNED_32_BIT_INTEGER);
     }
 
     @Override
